Single-source the per-backend crypto key internals
