+++ src/core/crypto/crypto_rsa_oaep_apple.cc
@@ -1,5 +1,7 @@
 #include <sourcemeta/core/crypto_rsa_oaep.h>
 
+#include "crypto_apple.h"
+
 #include <CoreFoundation/CoreFoundation.h> // CF*, kCF*
 #include <Security/Security.h>             // Sec*, kSec*
 
@@ -11,26 +13,6 @@
 #include <utility>     // std::unreachable
 
 namespace sourcemeta::core {
-
-// The layout matches the definitions in the sibling Apple key backends, since
-// each translation unit that reads a key redeclares its file-private members
-struct PublicKey::Internal {
-  PublicKey::Type kind;
-  SecKeyRef key;
-  std::string modulus;
-  std::size_t field_bytes;
-  std::string edwards_point;
-  EdwardsCurve edwards_curve;
-};
-
-struct PrivateKey::Internal {
-  PrivateKey::Type kind;
-  SecKeyRef key;
-  std::size_t field_bytes;
-  std::string edwards_seed;
-  EdwardsCurve edwards_curve;
-  bool rsa_pss_restricted{false};
-};
 
 namespace {
 auto make_data(const std::string_view value) -> CFDataRef {
