+++ src/core/crypto/crypto_sign_openssl.cc
@@ -21,21 +21,6 @@
 #include <string_view> // std::string_view
 #include <utility>     // std::move, std::unreachable
 
-namespace sourcemeta::core {
-
-// The parsed key keeps the native handle alive so that many signatures are
-// produced without rebuilding it
-struct PrivateKey::Internal {
-  PrivateKey::Type kind;
-  EVP_PKEY *key;
-  // The field width for the elliptic curve raw signature encoding
-  std::size_t field_bytes;
-  // Set for an id-RSASSA-PSS key, which is refused for PKCS1v15 signing
-  bool rsa_pss_restricted{false};
-};
-
-} // namespace sourcemeta::core
-
 namespace {
 
 auto to_message_digest(
