+++ src/core/crypto/crypto_verify_apple.cc
@@ -18,22 +18,6 @@
 #include <string_view> // std::string_view
 #include <utility>     // std::move, std::unreachable
 
-namespace sourcemeta::core {
-
-// The parsed key keeps the platform key object alive for reuse. The Edwards
-// curves have no Security framework primitive, so they keep the raw encoded
-// point and verify through CryptoKit or the reference implementation
-struct PublicKey::Internal {
-  PublicKey::Type kind;
-  SecKeyRef key;
-  std::string modulus;
-  std::size_t field_bytes;
-  std::string edwards_point;
-  EdwardsCurve edwards_curve;
-};
-
-} // namespace sourcemeta::core
-
 namespace {
 
 auto to_sec_key_pkcs1_v15_algorithm(
